Allow handle_user_stop = True when there are multiple WorkerManagers
The difference is that in the case of multiple WorkerManagers, those other than the first one will NOT execute the "countdown-and-sigkill" logic upon Ctrl+C. However, for all of the WMs, listening to and handling stop events will work normally.

Without this change, those other than the first WM will NOT respond to Ctrl+C correctly (e.g., not setting the stop event), which isn't the desired behavior.

PiperOrigin-RevId: 436695002
Change-Id: I11b50c7b06dbc49cac574978d535aca1259b5729