Fix launchpad tree map issue
`tree.map_structure` tries to create the original structure with the output of the map function, but here the map function returns only Nones, and so any object which cannot be constructed with None args will fail here.
We replace this `map_structure` with `traverse`, which more accurately reflects what we're trying to do.

PiperOrigin-RevId: 512583533
Change-Id: I77b56ae0c7fb5c3647d3f0f8820146b5bddc2887